feat(agent-email/npm): example smoke test with health-check diagnosis
Move scripts/demo.mjs -> examples/ (discoverable home for a runnable
example) and turn it into a one-command integration check: a health check
that names what's wrong (sidecar down / Lemonade not found / model not
downloaded) by diagnosing the triage probe's 502/timeout, then exercises
every standalone endpoint (health, version, emailHealth, emailVersion,
openapi, spec, draft, triage) with a PASS/SKIP tally. send() stays skipped
(connector-gated). Stays out of the npm tarball; linked from the hub.
